requirements.txt
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.1
opencv-python==4.8.0.74
pillow==9.5.0
scikit-image==0.20.0
scikit-learn==1.2.2
torch==2.0.1
torchvision==0.15.2
pandas==2.0.1
tqdm==4.65.0
jupyter==1.0.0
pytest==7.3.1
requests==2.31.0 